+++ CHANGES.md
@@ -10,9 +10,13 @@
 
 * add OpenTelemetry tracing to the FastAPI application
 
+### titiler.xarray
+
+* add `add_preview` in factory attribute (default to `False`)
+
 ### Misc
 
-* Add otel-collector and jaeger to the docker network 
+* Add otel-collector and jaeger to the docker network
 
 ## 0.22.4 (2025-07-02)